chat/requirements.txt
streamlit>=1.35.0
google-genai>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
